+++ packages/atomic/src/components/atomic-rating/atomic-rating.tsx
@@ -17,7 +17,7 @@ export const Rating: FunctionalComponent<RatingProps> = (props) => {
     return (
       <atomic-icon
         icon={props.icon}
-        class={active ? 'icon-active' : 'icon-inactive'}
+        class={`flex-shrink-0 ${active ? 'icon-active' : 'icon-inactive'}`}
         style={{width: iconSize, height: iconSize}}
       ></atomic-icon>
     );
@@ -45,7 +45,7 @@ export const Rating: FunctionalComponent<RatingProps> = (props) => {
         {emptyIconDisplay()}
       </div>
       <div
-        class="absolute left-0 top-0 z-10 flex items-center gap-0.5 pt-0.5 pb-0.5 overflow-hidden"
+        class="absolute left-0 top-0 z-1 flex items-center gap-0.5 pt-0.5 pb-0.5 overflow-hidden"
         style={{width}}
       >
         {filledIconDisplay()}