+++ packages/atomic/src/components/facets-v1/facet-show-more-less/facet-show-more-less.tsx
@@ -2,7 +2,7 @@ import {FunctionalComponent, h} from '@stencil/core';
 import {i18n} from 'i18next';
 import PlusIcon from 'coveo-styleguide/resources/icons/svg/plus.svg';
 import MinusIcon from 'coveo-styleguide/resources/icons/svg/minus.svg';
-import {createRipple} from '../../../utils/ripple';
+import {Button} from '../../common/button';
 
 interface FacetShowMoreProps {
   label: string;
@@ -25,17 +25,21 @@ export const FacetShowMoreLess: FunctionalComponent<FacetShowMoreProps> = (
   const showLessFacetValues = props.i18n.t('show-less-facet-values', {
     label,
   });
-  const btnClasses =
-    'flex items-baseline text-left p-2 text-sm mt-2 max-w-full text-primary rounded hover:text-primary focus:text-primary hover:bg-neutral-light focus:bg-neutral-light focus:outline-color';
+  const btnClasses = 'flex items-baseline text-left p-2 text-sm max-w-full';
   const iconClasses = 'w-2 h-2 mr-1';
-  return [
-    props.canShowLessValues && (
-      <button
+
+  if (!props.canShowLessValues && !props.canShowMoreValues) {
+    return;
+  }
+
+  return (
+    <div class="mt-2">
+      <Button
+        style="text-primary"
         part="show-less"
-        class={`show-less ${btnClasses}`}
+        class={`${btnClasses} ${props.canShowLessValues ? '' : 'hidden'}`}
         aria-label={showLessFacetValues}
         onClick={() => props.onShowLess()}
-        onMouseDown={(e) => createRipple(e, {color: 'neutral'})}
       >
         <atomic-icon
           part="show-more-less-icon"
@@ -43,15 +47,13 @@ export const FacetShowMoreLess: FunctionalComponent<FacetShowMoreProps> = (
           icon={MinusIcon}
         ></atomic-icon>
         <span class="truncate">{showLess}</span>
-      </button>
-    ),
-    props.canShowMoreValues && (
-      <button
+      </Button>
+      <Button
+        style="text-primary"
         part="show-more"
-        class={`show-more ${btnClasses}`}
+        class={`${btnClasses} ${props.canShowMoreValues ? '' : 'hidden'}`}
         aria-label={showMoreFacetValues}
         onClick={() => props.onShowMore()}
-        onMouseDown={(e) => createRipple(e, {color: 'neutral'})}
       >
         <atomic-icon
           part="show-more-less-icon"
@@ -59,7 +61,7 @@ export const FacetShowMoreLess: FunctionalComponent<FacetShowMoreProps> = (
           icon={PlusIcon}
         ></atomic-icon>
         <span class="truncate">{showMore}</span>
-      </button>
-    ),
-  ];
+      </Button>
+    </div>
+  );
 };