+++ packages/atomic/tailwind.config.js
@@ -40,13 +40,18 @@ module.exports = {
         sm: 'var(--atomic-text-sm)',
         base: 'var(--atomic-text-base)',
         lg: 'var(--atomic-text-lg)',
+        xl: 'var(--atomic-text-xl)',
+        '2xl': 'var(--atomic-text-2xl)',
       },
       screens: {
-        'desktop-only': {min: '768px'},
-        'mobile-only': {max: '767px'},
+        'desktop-only': {min: '1024px'},
+        'mobile-only': {raw: 'not all and (min-width: 1024px)'},
       },
       gridTemplateColumns: {
         'min-1fr': 'min-content 1fr',
+      },
+      zIndex: {
+        1: '1',
       },
     },
     backgroundColor: (theme) => ({