+++ packages/headless/src/state/state-sections.ts
@@ -27,6 +27,7 @@ import {FoldingState} from '../features/folding/folding-state';
 import {TriggerState} from '../features/triggers/triggers-state';
 import {QuestionAnsweringState} from '../features/question-answering/question-answering-state';
 import {StandaloneSearchBoxSetState} from '../features/standalone-search-box-set/standalone-search-box-set-state';
+import {RecentQueriesState} from '../features/recent-queries/recent-queries-state';
 
 export interface QuerySection {
   /**
@@ -253,4 +254,11 @@ export interface QuestionAnsweringSection {
    * The question and answers for a given query, also known as smart snippet.
    */
   questionAnswering: QuestionAnsweringState;
+}
+
+export interface RecentQueriesSection {
+  /**
+   * The recent queries executed by the current user.
+   */
+  recentQueries: RecentQueriesState;
 }