+++ packages/quantic/force-app/main/default/lwc/quanticFacet/quanticFacet.js
@@ -83,9 +83,7 @@ export default class QuanticFacet extends LightningElement {
   }
 
   disconnectedCallback() {
-    if (this.unsubscribe) {
-      this.unsubscribe();
-    }
+    this.unsubscribe?.();
   }
 
   updateState() {
@@ -94,11 +92,13 @@ export default class QuanticFacet extends LightningElement {
 
   get values() {
     return (
-      this.state.values.map((v) => ({
-        ...v,
-        checked: v.state === 'selected',
-        highlightedResult: v.value,
-      })) || []
+      this.state.values
+        .filter((value) => value.numberOfResults || value.state === 'selected')
+        .map((v) => ({
+          ...v,
+          checked: v.state === 'selected',
+          highlightedResult: v.value,
+        })) || []
     );
   }
 
@@ -107,7 +107,7 @@ export default class QuanticFacet extends LightningElement {
   }
 
   get canShowMoreSearchResults() {
-    return this.facet && this.facet.state.facetSearch.moreValuesAvailable;
+    return this.facet?.state.facetSearch.moreValuesAvailable;
   }
 
   get canShowMore() {