+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,101 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.51.1](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.51.0...@coveo/quantic@0.51.1) (2021-08-20)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.51.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.50.1...@coveo/quantic@0.51.0) (2021-08-20)
+
+
+### Features
+
+* **quantic:** standalone search box ([#1107](https://github.com/coveo/ui-kit/issues/1107)) ([88b26bb](https://github.com/coveo/ui-kit/commit/88b26bbf3a148cb77ce66cb0b1a7d5c9f77275bc))
+
+
+
+
+
+## [0.50.1](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.50.0...@coveo/quantic@0.50.1) (2021-08-19)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.50.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.49.2...@coveo/quantic@0.50.0) (2021-08-19)
+
+
+### Features
+
+* **quantic:** added optional chaining ([#1104](https://github.com/coveo/ui-kit/issues/1104)) ([d85c9e4](https://github.com/coveo/ui-kit/commit/d85c9e4e5fcde7255850b576350c6b5a6e02bc2f))
+
+
+
+
+
+## [0.49.2](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.49.1...@coveo/quantic@0.49.2) (2021-08-18)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [0.49.1](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.49.0...@coveo/quantic@0.49.1) (2021-08-18)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.49.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.48.1...@coveo/quantic@0.49.0) (2021-08-17)
+
+
+### Features
+
+* **quantic:** merge facet value components into one ([#1085](https://github.com/coveo/ui-kit/issues/1085)) ([6fa7fd0](https://github.com/coveo/ui-kit/commit/6fa7fd0964a18c50160daefd9713e9a48b0af8dc))
+
+
+
+
+
+## [0.48.1](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.48.0...@coveo/quantic@0.48.1) (2021-08-17)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.48.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.47.0...@coveo/quantic@0.48.0) (2021-08-16)
+
+
+### Features
+
+* **quantic:** add quantic-search-box-suggestions-list component ([#1071](https://github.com/coveo/ui-kit/issues/1071)) ([7179311](https://github.com/coveo/ui-kit/commit/717931144af95204d235fc471d285eb463fdd2bc))
+
+
+
+
+
+# [0.47.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.46.0...@coveo/quantic@0.47.0) (2021-08-16)
+
+
+### Features
+
+* **quantic:** SFINT-3987 Quantic QueryError component ([#990](https://github.com/coveo/ui-kit/issues/990)) ([43e3db2](https://github.com/coveo/ui-kit/commit/43e3db2731e496c0479c4679c3f088da3fc8cd02))
+
+
+
+
+
 # [0.46.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.45.0...@coveo/quantic@0.46.0) (2021-08-13)